txn: introduce interactive iproto transactions
With introduction of mvcc, both memtx and vinyl engines are
allowed to yield during a transaction. This makes possible to
implement interactive transactions: when txn statements don't
have to be sent at once.

Current implementation allows to have one transaction per connection.
More flexible approach is to multiplex transactions in one connection
using e.g. idea of streams. This is going to be implemented later.

Closes #2016