src/lua: fix running init lua script
When tarantool launched with -e flag
and in script after there is an error,
programm hangs. This happens because
shed fiber launches separate fiber for
init user script and starts auxiliary event
loop. It's supposed that fiber will stop this
loop, but in case of error in script, fiber tries
to stop a loop when the last one isn't started yet.

Added a flag, which will watch is loop started and
when fiber tries to call `ev_break()` we can be sure
that loop is runnig already.

Part of #4983