relay: rename is_raft_enabled message to relay_is_running
Relay sends a raft enabler/disabler message on start and stop. This
message would also be useful to perform some additional work related to relay
initialization/deinitialization done in the tx thread.

For example, the message will be used to notify tx thread that there's no
point in wating until a relay replicates a specific row, since the relay
is exiting. This is the case in #5435.

So rename the message and all the connected functions to represent its
enriched functionality.

Prerequisite #5435