test: make dict.items() compatible with Python 3.x
In Python 2.x calling keys makes a copy of the key that you can iterate over
while modifying the dict. This doesn't work in Python 3.x because keys returns
an iterator instead of a list. Python 3 raise an exception "dictionary changed
size during iteration". To workaround it one can use list to force a copy of
the keys to be made.

Part of #5538